Migrated from mysql database to sqlite

+++ sql/schema/20231025175954_recipes_owner_id.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 -- +goose StatementBegin
-ALTER TABLE recipes ADD COLUMN owner_id BIGINT NOT NULL DEFAULT -1;
+ALTER TABLE recipes ADD COLUMN owner_id INTEGER NOT NULL DEFAULT -1;
 -- +goose StatementEnd
 
 -- +goose Down
